新闻评论点赞/取消点赞功能开发

+++ utils/sliceUtils.go
@@ -1,7 +1,7 @@
 package utils
 
-func IsUintInSlice(slice []uint, item uint) bool {
-	m := make(map[uint]bool)
+func IsUintInSlice(slice []uint64, item uint64) bool {
+	m := make(map[uint64]bool)
 	for _, s := range slice {
 		m[s] = true
 	}
